AndresT_ParetoChart
This is a Pareto data chart that we created in IT 262. In this assignment we are calculating data for a series of items/factors over time using a Pareto Chart. In this example we are using the top reasons that compromise 80% of the phone calls to the IT help desk.  In Pareto original rule it states that 20% of issues are responsible for 80% complaints.