AndresT_AA Code
This is a code we did in IT 262 in class. In this chart we used a AA data code. This code was used to calculate the total score between A and B by using weight, criteria, and ratings in the data frame. Then it creates a chart with the results.